# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sefest Api Workshop",
        "description": "api ini hanya diperuntukan untuk peserta dan pemateri workshop sefest 2025",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "99539a8e18179644cbac3165cf994fe5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "email@sefest.com",
                                    "password": "PasswordSefest2025"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login Success"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|widjwijdiwjdialdjaiwjdiwjd"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "e62b6d5b548be003e473107717338fc7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "job": {
                                                "type": "string"
                                            },
                                            "bank_name": {
                                                "type": "string"
                                            },
                                            "account_number": {
                                                "type": "string"
                                            },
                                            "account_owner_name": {
                                                "type": "string"
                                            },
                                            "avatar": {
                                                "type": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "John",
                                    "email": "john@sefest.com",
                                    "password": "PasswordSefest2025",
                                    "job": "a student and freelancer full-stack engineer from indonesia, specialized in web development",
                                    "bank_name": "bca",
                                    "account_number": "1201220011",
                                    "account_owner_name": "johndoe",
                                    "avatar": "image.png"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "akun berhasil dibuat, silahkan ke halman login"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}